+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Add dark mode of Spinner component [#100](https://github.com/mishka-group/mishka_chelekom/pull/100)
 - Add dark mode of Banner component [#101](https://github.com/mishka-group/mishka_chelekom/pull/101)
 - Add dark mode of Breadcrumb component [#102](https://github.com/mishka-group/mishka_chelekom/pull/102)
+- Add dark mode of Blockquote component [#103](https://github.com/mishka-group/mishka_chelekom/pull/103)
 
 ### Bugs:
 - Fix un-CSP progress mounted [#72](https://github.com/mishka-group/mishka_chelekom/pull/72)